requirement.txt
streamlit
langchain
langchain-openai
faiss-cpu
sentence-transformers
pypdf
Pillow
rank-bm25
ragas
python-dotenv
